Update BitBucket Cloud to use API 2.0 URLs.
BitBucket Cloud 2FA does not work in VS for a few reasons. The first is
that Basic HTTP Auth is completely disabled when 2FA is enabled on the
account. Falling back to OAuth authentication works, but only for the
official REST API. Currently, BitBucket Source Link produces 'web
browser' URL's, using the
https://bitbucket.org/{user}/{repo}/raw/{commit}/{path} format. This is
not compatible with OAuth authentication.

By changing the URL's produced to the 2.0 REST API format, this should
allow VS to succesfully authenticate to private BitBucket Cloud Repo's
when using 2FA.