Fix packaging of Microsoft.Build.StandardCI (#1046)
When performing an official build, DebugType is portable which causes
the IncludeSymbols flag in src/Directory.Build.props to evaluate to
true. The Microsoft.Build.StandardCI project uses the
Microsoft.Build.NoTargets SDK which doesn't produce a build output (as
the package only ships a content file and not a binary). Change the
condition to check if the NoTargets SDK is used which indicates that a
symbol shouldn't be included.