Add RevisionTimestamp
Add a new MSBuild property called RevisionTimestamp that contains a
timestamp indicating when the latest commit was added to git. This can
then be used to timestamp the produced binaries (eg, dll files, file
creation times of the contents of nuget packages) to make .NET builds
and packages more deterministic/reproducible.

The timestamp is in the ISO 8601/RFC 3339 format, as generated by the
"O" formatter.

Fixes: #1450